Refactor EditUser command and DeleteUser command

+++ src/main/java/seedu/address/logic/commands/personcommands/DeleteUserCommand.java
@@ -1,16 +1,15 @@
 package seedu.address.logic.commands.personcommands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.core.Messages.MESSAGE_NO_USER_LOGGED_IN;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
 import java.util.List;
 
-import seedu.address.commons.core.Messages;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.exceptions.NoUserLoggedInException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
 
@@ -28,23 +27,18 @@ public class DeleteUserCommand extends Command {
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
 
-    //private final Index targetIndex;
-
     @Override
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
-        try {
-            requireNonNull(model);
-            List<Person> lastShownList = model.getFilteredPersonList();
-
-            Person personToDelete = model.getCurrentUser();
-
-            updateFriendListsDueToDeletedPerson(model, lastShownList, personToDelete);
-            model.removeCurrentUser();
-            model.deletePerson(personToDelete);
-            return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
-        } catch (NoUserLoggedInException e) {
-            throw new CommandException(Messages.MESSAGE_NO_USER_LOGGED_IN);
+        requireNonNull(model);
+        List<Person> lastShownList = model.getFilteredPersonList();
+        if (!model.hasSetCurrentUser()) {
+            throw new CommandException(MESSAGE_NO_USER_LOGGED_IN);
         }
+        Person personToDelete = model.getCurrentUser();
+        updateFriendListsDueToDeletedPerson(model, lastShownList, personToDelete);
+        model.removeCurrentUser();
+        model.deletePerson(personToDelete);
+        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
     }
 
     /**
